Upgrade to Next.js 15 with App Router
