+++ app/page.js
@@ -1,25 +1,18 @@
-import Head from "next/head";
 import Layout, { siteTitle } from "../componets/layout";
 import { getSortedPostsData } from "../lib/posts";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faGithub } from "@fortawesome/free-brands-svg-icons";
 import Link from "next/link";
 import Date from "../componets/date";
-export async function getStaticProps() {
-  const allPostsData = getSortedPostsData();
-  return {
-    props: {
-      allPostsData,
-    },
-  };
-}
 
-export default function Home({ allPostsData }) {
+export const metadata = {
+  title: siteTitle,
+};
+
+export default async function Home() {
+  const allPostsData = getSortedPostsData();
   return (
     <Layout home>
-      <Head>
-        <title>{siteTitle}</title>
-      </Head>
       <div className="sections">
         <h2 className="section-title">ハッカソン</h2>
 
@@ -232,7 +225,6 @@ export default function Home({ allPostsData }) {
             </li>))}
         </ul>
       </div>
-      <style jsx>{``}</style>
     </Layout>
   );
 }
