+++ package.json
@@ -19,9 +19,9 @@
     "@mui/styled-engine-sc": "^5.11.11",
     "date-fns": "^2.28.0",
     "gray-matter": "^4.0.3",
-    "next": "latest",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "next": "^15.3.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "remark": "^14.0.2",
     "remark-html": "^15.0.1",
     "styled-components": "^5.3.9"
